+++ frontend/src/app/index.js
@@ -15,10 +15,11 @@ import store from '../store';
 
 import NavBar from '../components/NavBar/NavBar';
 import Register from '../pages/Register';
-import Login from '../pages/Login';
+import Login from '../pages/Login/Login';
 import Home from '../pages/Home/Home';
 import PrivateRoute from '../components/private-route/PrivateRoute';
 import NewPage from '../components/NewPage/NewPage';
+import gray from '../images/gray.jpg';
 
 // Check for token to keep user logged in
 if (localStorage.jwtToken) {
@@ -40,23 +41,26 @@ if (localStorage.jwtToken) {
 }
 
 function App () {
-    return (
+    return ( 
         <Provider store={store}>
             <Router>
-                <div>
-                    <NavBar />
+                <div styles={{backgroundImage: `url(${gray})`}}>
+                  
+                    <div>
+                        <NavBar />
 
-                    <Switch>
-                        <Route
-                            exact
-                            path='/'
-                            component={() => <Redirect to='/login' />}
-                        />
-                        <Route exact path='/newpage' component={NewPage}/>
-                        <Route exact path='/login' component={Login} />
-                        <Route exact path='/register' component={Register} />
-                        <PrivateRoute exact path='/home' component={Home} />
-                    </Switch>
+                        <Switch>
+                            <Route
+                                exact
+                                path='/'
+                                component={() => <Redirect to='/login' />}
+                            />
+                            <Route exact path='/newpage' component={NewPage}/>
+                            <Route exact path='/login' component={Login} />
+                            <Route exact path='/register' component={Register} />
+                            <PrivateRoute exact path='/home' component={Home} />
+                        </Switch>
+                        </div>
                 </div>
             </Router>
         </Provider>
