+++ frontend/src/pages/Login/Login.jsx
@@ -1,0 +1,108 @@
+import React, { Component } from "react";
+import { Link } from "react-router-dom";
+import PropTypes from "prop-types";
+import { connect } from "react-redux";
+import { loginUser } from "../../actions/authActions";
+import { Button, InputLabel , Input , LoginContainer} from './Login.styled'
+
+class Login extends Component {
+    constructor() {
+        super();
+        this.state = {
+            email: "",
+            password: "",
+            errors: {},
+        };
+    }
+
+    componentWillReceiveProps(nextProps) {
+        if (nextProps.auth.isAuthenticated) {
+            this.props.history.push("/home");
+        }
+
+        if (nextProps.errors) {
+            this.setState({
+                errors: nextProps.errors,
+            });
+        }
+    }
+
+    onChange = (e) => {
+        this.setState({ [e.target.id]: e.target.value });
+    };
+
+    onSubmit = (e) => {
+        e.preventDefault();
+
+        const userData = {
+            email: this.state.email,
+            password: this.state.password,
+        };
+
+        this.props.loginUser(userData);
+    };
+
+    render() {
+        const { errors } = this.state;
+
+        return (
+            <LoginContainer>
+         
+                <div>
+                    <h4>
+                        <InputLabel>Sign In!</InputLabel>
+                    </h4>
+
+                    <form noValidate onSubmit={this.onSubmit}>
+                        <div>
+                            
+                            <InputLabel htmlFor="email">Email</InputLabel>
+                            <span className="red-text">
+                                {errors.email}
+                                {errors.emailnotfound}
+                            </span>
+                            <Input
+                                onChange={this.onChange}
+                                value={this.state.email}
+                                error={errors.email}
+                                id="email"
+                                type="email"
+                            />
+                        </div>
+                        <div>
+                            
+                            <InputLabel htmlFor="password">Password</InputLabel>
+                            <span className="red-text">
+                                {errors.password}
+                                {errors.passwordincorrect}
+                            </span>
+                            <Input
+                                onChange={this.onChange}
+                                value={this.state.password}
+                                error={errors.password}
+                                id="password"
+                                type="password"
+                            />
+                        </div>
+                        <div>
+                            <Button type="submit">Login</Button>
+                        </div>
+                    </form>
+                </div>
+            </LoginContainer>
+        );
+    }
+}
+
+Login.propTypes = {
+    loginUser: PropTypes.func.isRequired,
+    auth: PropTypes.object.isRequired,
+    errors: PropTypes.object.isRequired,
+};
+
+const mapStateToProps = (state) => ({
+    auth: state.auth,
+    errors: state.errors,
+});
+
+export default connect(mapStateToProps, { loginUser })(Login);
